moved time stores outside the cache module
